refactor(circle): decompose enums and classes into multiple files
Good practice
