+++ lib/circle/enums/arc_type.dart
@@ -1,0 +1,4 @@
+enum ArcType {
+  HALF,
+  FULL,
+}
