+++ lib/circle/enums/circular_stroke_cap.dart
@@ -1,0 +1,1 @@
+enum CircularStrokeCap { butt, round, square }
